Refactored state into separate modules
