+++ index.html
@@ -1,8 +1,8 @@
----
+<!-- ---
 title: Crafting Conundrums
 book: 'Crafting Conundrums: Puzzles and Patterns for the Bead Crochet Artist'
 description: Design Tool accompanying the book
----
+--- -->
 <html>
   <head>
     <link rel="stylesheet" type="text/css" href="css/beadCrochetApp.css">
