+++ js/images.js
@@ -18,39 +18,39 @@ export const generateImageBlobFromSVG = ( svgText, newWidth, mimeType, quality )
 
   return new Promise( (resolve, reject) =>
   {
-      // 1. Create an abstract canvas
-      let canvas = document.createElement('canvas');
-      let ctx = canvas.getContext("2d");
+    // 1. Create an abstract canvas
+    let canvas = document.createElement('canvas');
+    let ctx = canvas.getContext("2d");
 
-      // 2. Create an image element to load the SVG
-      let img = new Image();
+    // 2. Create an image element to load the SVG
+    let img = new Image();
 
-      // 3. Manipulate
-      img.onload = function() {
-          // Declare initial dimensions of the image
-          let originalWidth = img.width;
-          let originalHeight = img.height;
+    // 3. Manipulate
+    img.onload = function() {
+        // Declare initial dimensions of the image
+        let originalWidth = img.width;
+        let originalHeight = img.height;
 
-          // Declare the new width of the image
-          // And calculate the new height to preserve the aspect ratio
-          img.width = newWidth;
-          img.height = (originalHeight / originalWidth) * newWidth;
+        // Declare the new width of the image
+        // And calculate the new height to preserve the aspect ratio
+        img.width = newWidth;
+        img.height = (originalHeight / originalWidth) * newWidth;
 
-          // Set the dimensions of the canvas to the new dimensions of the image
-          canvas.width = img.width;
-          canvas.height = img.height;
+        // Set the dimensions of the canvas to the new dimensions of the image
+        canvas.width = img.width;
+        canvas.height = img.height;
 
-          // Render image in Canvas
-          ctx.drawImage(img, 0, 0, img.width, img.height); 
+        // Render image in Canvas
+        ctx.drawImage(img, 0, 0, img.width, img.height); 
 
-          // Export the canvas to blob
-          // You may modify this to export it as a base64 data URL
-          canvas.toBlob(function(blob){
-              resolve(blob);
-          }, mimeType, quality);
-      };
+        // Export the canvas to blob
+        // You may modify this to export it as a base64 data URL
+        canvas.toBlob(function(blob){
+            resolve(blob);
+        }, mimeType, quality);
+    };
 
-      // Load the DataURL of the SVG
-      img.src = dataURL;
-    });
-  }
+    // Load the DataURL of the SVG
+    img.src = dataURL;
+  });
+}
